requirements.txt
PyQt5
qrcode
Pillow
